We need those system to work without importing 3d party libraries
