+++ napalm_base/helpers.py
@@ -31,22 +31,6 @@ _MACFormat.word_fmt = '%.2X'
 # ----------------------------------------------------------------------------------------------------------------------
 # callable helpers
 # ----------------------------------------------------------------------------------------------------------------------
-def find_version(*file_paths):
-    """
-    This pattern was modeled on a method from the Python Packaging User Guide:
-        https://packaging.python.org/en/latest/single_source_version.html
-
-    We read instead of importing so we don't get import errors if our code
-    imports from dependencies listed in install_requires.
-    """
-    base_module_file = os.path.join(*file_paths)
-    with open(base_module_file) as f:
-        base_module_data = f.read()
-    version_match = re.search(r"^__version__ = ['\"]([^'\"]*)['\"]",
-                              base_module_data, re.M)
-    if version_match:
-        return version_match.group(1)
-    raise RuntimeError("Unable to find version string.")
 
 def load_template(cls, template_name, template_source=None, template_path=None,
                   openconfig=False, **template_vars):
